refactor(backend): extract embedding crud module
Pure move: all 8 embedding functions come out of crud/__init__.py into
crud/embedding.py with byte-identical bodies, verified by parsing both files with
ast and comparing the unparsed functions.

No transitional re-export at the extraction site -- the 11 references across 4
files are migrated in the same commit, per the AGENTS.md rule against reaching
through the crud package root.

graph_builder.py loses its `crud` import entirely (that call was its only use);
generator.py and services/explorer/embeddings.py keep theirs for get_model and
get_user_by_id. crud/explorer.py splits its one import in two, leaving the
existing comment attached to the _TEST_SET_RELATED_FIELDS line it explains.

The function-local imports of EmbeddingStatus and Status inside
get_active_embeddings_for_entities stay function-local, to keep the move pure.

No import in crud/__init__.py became unused -- UUID, QueryBuilder and the
crud_utils helpers are all still used by the blocks that remain.
